Consolidate mcp_utils.rs
  - DEFAULT_MAX_ITERATIONS - constant
  - McpLoopConfig - config struct
  - extract_server_label() - helper function
  - ensure_request_mcp_client() - MCP connection function
