+++ sgl-model-gateway/src/routers/openai/router.rs
@@ -23,7 +23,7 @@ use super::{
         ComponentRefs, PayloadState, RequestContext, ResponsesComponents, SharedComponents,
         WorkerSelection,
     },
-    mcp::{execute_tool_loop, prepare_mcp_payload_for_streaming, McpLoopConfig},
+    mcp::{execute_tool_loop, prepare_mcp_payload_for_streaming},
     provider::ProviderRegistry,
     responses::{mask_tools_as_mcp, patch_streaming_response_json},
     streaming::handle_streaming_response,
@@ -47,7 +47,7 @@ use crate::{
     routers::{
         conversations::persist_conversation_items,
         header_utils::{apply_provider_headers, extract_auth_header},
-        mcp::ensure_request_mcp_client,
+        mcp_utils::{ensure_request_mcp_client, McpLoopConfig},
     },
 };
